refactor runtime-api to a standalone crate, to allows use of Color etc. in pax-manifest without circular dependencies
