+++ pax-compiler/src/helpers.rs
@@ -33,7 +33,7 @@ pub const ERR_SPAWN: &str = "failed to spawn child";
 
 //whitelist of package ids that are relevant to the compiler, e.g. for cloning & patching, for assembling FS paths,
 //or for looking up package IDs from a userland Cargo.lock.
-pub const ALL_PKGS: [&'static str; 13] = [
+pub const ALL_PKGS: [&'static str; 14] = [
     "pax-cartridge",
     "pax-chassis-common",
     "pax-chassis-ios",
@@ -42,6 +42,7 @@ pub const ALL_PKGS: [&'static str; 13] = [
     "pax-cli",
     "pax-compiler",
     "pax-runtime",
+    "pax-runtime-api",
     "pax-engine",
     "pax-macro",
     "pax-message",
