+++ pax-engine/src/lib.rs
@@ -5,7 +5,7 @@ pub use log;
 pub use pax_runtime::api;
 pub use pax_runtime::engine::node_interface::*;
 pub use pax_runtime::layout;
-pub use pax_runtime::math;
+pub use pax_runtime::api::math;
 pub use pax_runtime::rendering;
 
 pub use pax_runtime::api::serde;
