+++ pax-manifest/Cargo.toml
@@ -16,6 +16,7 @@ parsing = ["pest", "pest_derive"]
 [dependencies]
 serde = { version="1.0.95", features=["derive"]}
 pax-message = {path="../pax-message", version="0.12.3"}
+pax-runtime-api = {path="../pax-runtime-api", version="0.12.3"}
 pest = { version="2.7.6", optional=true}
 pest_derive = {version="2.7.6", optional=true}
 serde_with = { version= "3.6.1", features = ["json"]}
