+++ pax-runtime-api/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "pax-runtime-api"
+version = "0.12.3"
+edition = "2021"
+authors = ["Zack Brown <zack@pax.dev>", "Warfa Jibril <warfa@pax.dev>", "Samuel Selleck <samuel@pax.dev>"]
+license = "MIT OR Apache-2.0"
+homepage = "https://pax.dev/"
+repository = "https://www.github.com/paxproject/pax"
+description = "Userland constructs used at the runtime API boundary of Pax Engine"
+
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = {version = "1.0.159", features=["derive"] }
+wasm-bindgen = {version = "0.2.80", features=["serde-serialize"]}
+kurbo = "0.9.0"
+piet = "0.6.0"
+pax-message = {version="0.12.3", path="../pax-message"}
