+++ pax-runtime/src/lib.rs
@@ -6,10 +6,8 @@ pub mod conditional;
 pub mod constants;
 pub mod declarative_macros;
 pub mod engine;
-pub mod expressions;
 pub mod form_event;
 pub mod layout;
-pub mod math;
 pub mod numeric;
 pub mod properties;
 pub mod rendering;
@@ -19,7 +17,6 @@ pub mod slot;
 pub use crate::component::*;
 pub use crate::conditional::*;
 pub use crate::engine::*;
-pub use crate::expressions::*;
 pub use crate::layout::*;
 pub use crate::properties::*;
 pub use crate::rendering::*;
