+++ pax-runtime/src/rendering.rs
@@ -1,12 +1,12 @@
 use std::any::Any;
 use std::cell::RefCell;
 
-use super::math::Point2;
+use super::api::math::Point2;
 use std::iter;
 use std::rc::Rc;
 
 use crate::api::{CommonProperties, RenderContext, Window};
-use crate::math::Transform2;
+use crate::api::math::Transform2;
 use crate::node_interface::NodeLocal;
 use pax_manifest::UniqueTemplateNodeIdentifier;
 use piet::{Color, StrokeStyle};
